fix: support age as a float
The new version of BIDS validator, at 2.0.6,
returns `age` in `SubjectMetadata` as a float

+++ dandi/bids_validator_deno/_models.py
@@ -72,5 +72,5 @@ class SummaryOutput(_BaseModel):
 
 class SubjectMetadata(_BaseModel):
     participantId: str
-    age: Union[int, Literal["89+"], None] = None
+    age: Union[float, Literal["89+"], None] = None
     sex: Optional[str] = None
